Add support for seqs of nodes (from a map call or something)
Could probably optimise this so that the flattening only needs to happen once,
but, meh. It only flattens if that level needs checking. If we don't recur into
children then it doesn't bother which is nice.

I think this way is potentially slower in some cases, but leaves less questions
around "will this matter in some comparison at some point". Like, technically
flattening then storing that flattened version actually deleted some information
about the structure of the tree.

That could be important information, so I'll take the simple and safe route.

Future people and / or me, please don't be mad.