Broke tests, but started lifecycle hooks, kinda don't like the name
It's too Reacty, maybe I can get away without them for now? Or come up with a
better way to represent this. Also need to improve the spec around attrs so that
I don't need to do this extra parsing in the attrs fn. The only auxiliary fns I
should need should be doing simple extraction on the pre-parsed tree,
not transformations.