Made sure I don't rely on (:state opts) in event listeners
The (:host opts) is the only thing we can really assume won't change. If that
changes we're on another tree so who cares. The important thing is that all
lookups of state are performed against the top level host node. The state is
updated on that node on every render.

It may be nil.