Refactor out :state, replace with :state!
So now to the user all they have is a state value (ideally a map) plz.

As soon as it's inside the system it is wrapped up in an atom, every time it
leaves and the user sees it, it is unwrapped. The use of an atom should be
invisible to the user (as long as they don't use internal functions *GLARE*).