Added test that reveals issues with my approach
A change further down in the diff can point to a different node than intended
because of the way indexes work. Maybe the most elegant thing will be to replace
nil nodes with comment nodes or something. That would keep the indexing
consistent.

That may still fall down and remove the wrong nodes at some points though. Hmm.